Introduce middleware via system so we can declare
different middleware stacks for different environments. For example,
certain security or cross-origin constraints meant for production make
it hard to develop and test locally. So it's useful to be able to
define a local dev/test-specific middleware stack.

Middleware is wrapped over the root handler in the order of index
position of middleware. For example:

    (def dummy-handler identity)

    ;; Middleware must be applied in correct order of dependency
    (defn wrap-inc [h] (fn [req] (inc (h req))))
    (defn wrap-str [h] (fn [req] (str (h req))))
    (defn wrap-vec [h] (fn [req] (vec (h req))))

    (let [composed-middleware
          (apply comp
                 (reverse [wrap-inc    ; 0th middleware wrapped first
                           wrap-str    ; 1st middleware wrapped next
                           wrap-vec])) ; 2nd middleware wrapped last

          wrapped-handler
          (composed-middleware dummy-handler)]
      (wrapped-handler 42)) ; => [\4 \3]