Add hiccup layout, use existing :application/view param
to drive the rendering logic via middleware, staying true to the
design choice made in seancorfield/usermanager-example. Except, we use
hiccup instead of Selmer for templating.

Now, the main app renders an HTML page, when a request matches an
existing route, because we have configured it with the wrap-page
middleware. When routes don't match, we render a 404 page.

Extant tests continue working as-is, because we have not added
middleware to the test middleware stack. If we want to test layout
paths, we can add handler tests to check the hiccup data returned.

Styling is in-line because we don't yet have a resource route, with
which we can return static assets such as a CSS file.