Add empty dev directory so people don't trip on
the "SQL error or missing database" error.

ref github issue:

https://github.com/adityaathalye/usermanager-first-principles/issues/1

 tried to run it empty, got missing table error #1

    SLF4J: Defaulting to no-operation (NOP) logger implementation
    SLF4J: See https://www.slf4j.org/codes.html#noProviders for further details.
    Setting up DB having dbname:  dev/usermanager_dev_db.sqlite3
    Created database and addressbook table!
    Exception: [SQLITE_ERROR] SQL error or missing database (no such table: department)
    Unable to populate the initial data -- proceed with caution!
    Starting up Server on port:  3000

The program expects a `dev` directory, to be able to create the dbname
which includes the dev path (printed out at startup). If the directory
is absent, the SQLITE_ERROR occurs. Manually creating the directory
fixes the problem, but it's better I just commit the empty dir to
source to remove this point of friction.

Properly-speaking, the startup process should die if it can't find the
DB (or if any of the setup goes bad), but I don't want to add too much
error handling in this project as it will obscure the essential
details.